Populate account_address with past and future HD Wallet addresses.
The account_address table associates dogecoin addresses with an Account
so we can find the owner account of UTXOs as we process new blocks.

This is necessary so we can spend those UTXOs later, and recognise
payments toward Invoices.

Effectively the account_address table contains a pool of both used and
not-yet-used HD Wallet addresses for every account in Gigawallet.

There are a few other fixes, e.g. tracking whether a KeyIndex is an
internal (change) or external (invoice) address, removed the old
GetPendingInvoices stub.