Split the API between two ports, admin and private
The purpose here is to allow firewalling of the admin API while
public proxying to the public api.

The connect & qr endpoints only work on the public API server now.

Introduced new config for different webservers, as well as a new
entry for the public facing root URL so that the QR code can point
to the correct dogecoin connect endpoint.