Fallback technique for AuxPoW block parsing, with tests
In case the block parser fails on an AuxPoW block, chainfollower
will use an alternate method where it fetches the raw header from Core,
strips off any additional (AuxPoW) data after the standard 80-byte
header, and parses the block using only that 80-byte header.
Warnings are logged if this happens.

This is similar to the way libdogecoin parses blocks (it parses the
header first, then parses the block data separately)